Added the ability to edit map objects.  Next will come the ability to create, delete and move map objects and pages within them.

+++ trunk/turbu/engines/map engine/tiles.pas
@@ -33,7 +33,7 @@ type
    TTile = class abstract(TParentSprite)
    protected
    class var
-      FHeartbeat: byte;
+      FHeartbeat: integer;
       FBroadcastList: TList<TBroadcastProc>;
       class constructor Create;
       class destructor Destroy;
